+++ Assets/Scenes/Main.unity
@@ -129,24 +129,6 @@ Transform:
     type: 3}
   m_PrefabInstance: {fileID: 2207445858600062310}
   m_PrefabAsset: {fileID: 0}
---- !u!4 &322115224 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 5622831708247501682, guid: 32c61e679e516144fa6b2a55a003fb44,
-    type: 3}
-  m_PrefabInstance: {fileID: 5622831708531736042}
-  m_PrefabAsset: {fileID: 0}
---- !u!114 &322115225 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 5622831708247501683, guid: 32c61e679e516144fa6b2a55a003fb44,
-    type: 3}
-  m_PrefabInstance: {fileID: 5622831708531736042}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 253f3e33935bb1741ac63fd446049ca2, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &330585543
 GameObject:
   m_ObjectHideFlags: 0
@@ -865,8 +847,6 @@ Transform:
   - {fileID: 1211926612}
   - {fileID: 1876077630}
   - {fileID: 1834058038}
-  - {fileID: 322115224}
-  - {fileID: 3116618265053219332}
   - {fileID: 7976951904415384665}
   - {fileID: 2036168304}
   m_Father: {fileID: 0}
@@ -1023,6 +1003,7 @@ Transform:
   - {fileID: 1184505000}
   - {fileID: 1791522741}
   - {fileID: 59282081}
+  - {fileID: 3116618265053219332}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -1209,10 +1190,8 @@ MonoBehaviour:
   levelLoader: {fileID: 1211926613}
   sfxManager: {fileID: 1876077631}
   musicManager: {fileID: 1834058039}
-  currentProgressManager: {fileID: 322115225}
   mainMenu: {fileID: 2036168305}
   pauseMenu: {fileID: 7976951904415384667}
-  countersUI: {fileID: 1106137028625975699}
   isOnMainMenu: 0
 --- !u!1001 &2036168303
 PrefabInstance:
@@ -1244,7 +1223,7 @@ PrefabInstance:
     - target: {fileID: 4879172274831871519, guid: efa981397779bc049ae0dfcbeba7dfb1,
         type: 3}
       propertyPath: m_RootOrder
-      value: 7
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 4879172274831871519, guid: efa981397779bc049ae0dfcbeba7dfb1,
         type: 3}
@@ -1367,7 +1346,7 @@ PrefabInstance:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Modification:
-    m_TransformParent: {fileID: 862587028}
+    m_TransformParent: {fileID: 1385518682}
     m_Modifications:
     - target: {fileID: 2601314807413310357, guid: b6ef121d10cc84d48abd701b8621f0bf,
         type: 3}
@@ -1387,7 +1366,7 @@ PrefabInstance:
     - target: {fileID: 2601314807413310358, guid: b6ef121d10cc84d48abd701b8621f0bf,
         type: 3}
       propertyPath: m_RootOrder
-      value: 5
+      value: 4
       objectReference: {fileID: 0}
     - target: {fileID: 2601314807413310358, guid: b6ef121d10cc84d48abd701b8621f0bf,
         type: 3}
@@ -1479,20 +1458,13 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 7060435805179464277, guid: b6ef121d10cc84d48abd701b8621f0bf,
+        type: 3}
+      propertyPath: slapManager
+      value: 
+      objectReference: {fileID: 2207445858600062311}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: b6ef121d10cc84d48abd701b8621f0bf, type: 3}
---- !u!114 &1106137028625975699 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 7060435805179464277, guid: b6ef121d10cc84d48abd701b8621f0bf,
-    type: 3}
-  m_PrefabInstance: {fileID: 1106137028625975698}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6480b1688c65b9e4b801ce98024d1651, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1001 &2207445858600062310
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1562,6 +1534,18 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: a24288dddbba9cc418c16a495665372d, type: 3}
+--- !u!114 &2207445858600062311 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 2207445858575392708, guid: a24288dddbba9cc418c16a495665372d,
+    type: 3}
+  m_PrefabInstance: {fileID: 2207445858600062310}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b98fdd8b4f02171439a0ac09df844c30, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &3069713098190413128
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1662,7 +1646,7 @@ PrefabInstance:
     - target: {fileID: 4801561748260402798, guid: 6d7e4ce42b193814a9b20df1b540a648,
         type: 3}
       propertyPath: m_RootOrder
-      value: 6
+      value: 4
       objectReference: {fileID: 0}
     - target: {fileID: 4801561748260402798, guid: 6d7e4ce42b193814a9b20df1b540a648,
         type: 3}
@@ -1894,75 +1878,6 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: d4496018b327dfb47b7454b840c8ce37, type: 3}
---- !u!1001 &5622831708531736042
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 862587028}
-    m_Modifications:
-    - target: {fileID: 5622831708247501682, guid: 32c61e679e516144fa6b2a55a003fb44,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 4
-      objectReference: {fileID: 0}
-    - target: {fileID: 5622831708247501682, guid: 32c61e679e516144fa6b2a55a003fb44,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 958.82416
-      objectReference: {fileID: 0}
-    - target: {fileID: 5622831708247501682, guid: 32c61e679e516144fa6b2a55a003fb44,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 538.1785
-      objectReference: {fileID: 0}
-    - target: {fileID: 5622831708247501682, guid: 32c61e679e516144fa6b2a55a003fb44,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0.43372834
-      objectReference: {fileID: 0}
-    - target: {fileID: 5622831708247501682, guid: 32c61e679e516144fa6b2a55a003fb44,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 5622831708247501682, guid: 32c61e679e516144fa6b2a55a003fb44,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5622831708247501682, guid: 32c61e679e516144fa6b2a55a003fb44,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5622831708247501682, guid: 32c61e679e516144fa6b2a55a003fb44,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5622831708247501682, guid: 32c61e679e516144fa6b2a55a003fb44,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5622831708247501682, guid: 32c61e679e516144fa6b2a55a003fb44,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5622831708247501682, guid: 32c61e679e516144fa6b2a55a003fb44,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5622831708247501693, guid: 32c61e679e516144fa6b2a55a003fb44,
-        type: 3}
-      propertyPath: m_Name
-      value: CurrentProgressManager
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 32c61e679e516144fa6b2a55a003fb44, type: 3}
 --- !u!224 &7976951904415384665 stripped
 RectTransform:
   m_CorrespondingSourceObject: {fileID: 4801561748260402798, guid: 6d7e4ce42b193814a9b20df1b540a648,
