--- Assets/Scripts/Managers/CurrentProgressManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 253f3e33935bb1741ac63fd446049ca2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
